rebased so as to remove multiple empty overrides
